pyproject.toml
[project]
name = "lorscan"
version = "0.1.0"
description = "Local Lorcana collection manager with Claude-vision card recognition"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "Pillow>=11.0",
]

[project.scripts]
lorscan = "lorscan.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "freezegun>=1.5",
    "ruff>=0.7",
    "syrupy>=4.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "live: hits real external APIs; requires --runlive flag",
]
asyncio_mode = "auto"
